+++ app/src/main/java/com/battlelancer/seriesguide/shows/database/SgSeason2Helper.kt
@@ -1,9 +1,8 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2021-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.shows.database
 
-import androidx.lifecycle.LiveData
 import androidx.room.ColumnInfo
 import androidx.room.Dao
 import androidx.room.Insert
@@ -12,6 +11,7 @@ import androidx.room.Transaction
 import androidx.room.Update
 import com.battlelancer.seriesguide.provider.SeriesGuideContract
 import com.battlelancer.seriesguide.provider.SeriesGuideContract.SgSeason2Columns
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface SgSeason2Helper {
@@ -66,22 +66,19 @@ interface SgSeason2Helper {
      * Excludes seasons where total episode count is 0.
      */
     @Query("SELECT * FROM sg_season WHERE series_id = :showId AND season_totalcount != 0 ORDER BY season_number DESC")
-    fun getSeasonsOfShowLatestFirst(showId: Long): LiveData<List<SgSeason2>>
+    fun getSeasonsOfShowLatestFirst(showId: Long): Flow<List<SgSeason2>>
 
     /**
      * Excludes seasons where total episode count is 0.
      */
     @Query("SELECT * FROM sg_season WHERE series_id = :showId AND season_totalcount != 0 ORDER BY season_number ASC")
-    fun getSeasonsOfShowOldestFirst(showId: Long): LiveData<List<SgSeason2>>
+    fun getSeasonsOfShowOldestFirst(showId: Long): Flow<List<SgSeason2>>
 
     /**
      * Note: does not exclude seasons based on season_totalcount because it might not be up-to-date.
      */
     @Query("SELECT * FROM sg_season WHERE series_id = :showId ORDER BY season_number ASC")
     fun getSeasonsForExport(showId: Long): List<SgSeason2>
-
-    @Update(entity = SgSeason2::class)
-    fun updateSeasonCounters(seasonCountUpdate: SgSeason2CountUpdate)
 
     @Query("DELETE FROM sg_season WHERE series_id = :showId")
     suspend fun deleteSeasonsOfShow(showId: Long): Int
@@ -97,15 +94,6 @@ data class SgSeason2Numbers(
     val number: Int
         get() = numberOrNull ?: 0 // == Specials, but should ignore seasons without number.
 }
-
-data class SgSeason2CountUpdate(
-    @ColumnInfo(name = SgSeason2Columns._ID) val id: Long,
-    @ColumnInfo(name = SgSeason2Columns.WATCHCOUNT) val notWatchedReleasedCount: Int,
-    @ColumnInfo(name = SgSeason2Columns.UNAIREDCOUNT) val notWatchedToBeReleasedCount: Int,
-    @ColumnInfo(name = SgSeason2Columns.NOAIRDATECOUNT) val notWatchedNoReleaseCount: Int,
-    @ColumnInfo(name = SgSeason2Columns.TOTALCOUNT) val totalCount: Int,
-    @ColumnInfo(name = SgSeason2Columns.TAGS) val tags: String
-)
 
 data class SgSeason2Update(
     @ColumnInfo(name = SgSeason2Columns._ID) val id: Long,