+++ app/src/main/java/com/battlelancer/seriesguide/shows/overview/SeasonsAdapter.kt
@@ -1,5 +1,5 @@
-// Copyright 2023 Uwe Trottmann
 // SPDX-License-Identifier: Apache-2.0
+// Copyright 2021-2024 Uwe Trottmann
 
 package com.battlelancer.seriesguide.shows.overview
 
@@ -15,6 +15,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.battlelancer.seriesguide.R
 import com.battlelancer.seriesguide.databinding.ItemSeasonBinding
 import com.battlelancer.seriesguide.shows.database.SgSeason2
+import com.battlelancer.seriesguide.shows.overview.SeasonsViewModel.SgSeasonWithStats
 import com.battlelancer.seriesguide.util.TextTools
 import com.uwetrottmann.androidutils.AndroidUtils
 
@@ -24,7 +25,7 @@ import com.uwetrottmann.androidutils.AndroidUtils
 class SeasonsAdapter(
     private val context: Context,
     private val itemClickListener: ItemClickListener
-) : ListAdapter<SgSeason2, SeasonsAdapter.ViewHolder>(SgSeason2DiffCallback()) {
+) : ListAdapter<SgSeasonWithStats, SeasonsAdapter.ViewHolder>(SgSeason2DiffCallback()) {
 
     interface ItemClickListener {
         fun onItemClick(v: View, seasonRowId: Long)
@@ -52,19 +53,20 @@ class SeasonsAdapter(
             }
         }
 
-        fun bindTo(season: SgSeason2, context: Context) {
-            this.season = season
+        fun bindTo(seasonWithStats: SgSeasonWithStats, context: Context) {
+            val season = seasonWithStats.season.also { this.season = it }
+            val stats = seasonWithStats.stats
 
             // Title
             binding.textViewSeasonTitle.text = SeasonTools.getSeasonString(context, season.number)
 
             // Not watched episodes by type.
-            val released = season.notWatchedReleased
-            val toBeReleased = season.notWatchedToBeReleased
-            val noRelease = season.notWatchedNoRelease
+            val released = stats.notWatchedReleased
+            val toBeReleased = stats.notWatchedToBeReleased
+            val noRelease = stats.notWatchedNoRelease
 
             // Progress bar
-            val max = season.total
+            val max = stats.total
             val progress = max - released - toBeReleased - noRelease
             binding.progressBarSeason.apply {
                 this.max = max
@@ -82,8 +84,11 @@ class SeasonsAdapter(
                 res.getString(R.string.format_progress_and_total, progress, max)
             }
 
-            // Skipped indicator
-            binding.imageViewSeasonSkipped.isGone = !SeasonTools.hasSkippedTag(season.tags)
+            // Skipped and collected indicator
+            val skipped = stats.skipped
+            val collected = stats.collected
+            binding.imageViewSeasonSkipped.isGone = skipped == 0
+            binding.imageViewSeasonCollected.isGone = collected == 0
 
             // Status text
             val countText = StringBuilder()
@@ -128,10 +133,25 @@ class SeasonsAdapter(
                     )
                 )
             }
+            if (skipped > 0) {
+                if (countText.isNotEmpty()) countText.append(" · ")
+                countText.append(
+                    res.getQuantityString(
+                        R.plurals.skipped_episodes_plural,
+                        skipped, skipped
+                    )
+                )
+            }
+            if (collected > 0) {
+                if (countText.isNotEmpty()) countText.append(" · ")
+                countText.append(
+                    res.getQuantityString(
+                        R.plurals.collected_episodes_plural,
+                        collected, collected
+                    )
+                )
+            }
             binding.textViewSeasonWatchCount.text = countText
-
-            // Context menu
-
         }
 
         companion object {
@@ -157,10 +177,12 @@ class SeasonsAdapter(
 
 }
 
-class SgSeason2DiffCallback : DiffUtil.ItemCallback<SgSeason2>() {
-    override fun areItemsTheSame(oldItem: SgSeason2, newItem: SgSeason2): Boolean =
-        oldItem.id == newItem.id
+class SgSeason2DiffCallback : DiffUtil.ItemCallback<SgSeasonWithStats>() {
+    override fun areItemsTheSame(oldItem: SgSeasonWithStats, newItem: SgSeasonWithStats): Boolean =
+        oldItem.season.id == newItem.season.id
 
-    override fun areContentsTheSame(oldItem: SgSeason2, newItem: SgSeason2): Boolean =
-        oldItem == newItem
+    override fun areContentsTheSame(
+        oldItem: SgSeasonWithStats,
+        newItem: SgSeasonWithStats
+    ): Boolean = oldItem == newItem
 }